Time: 0 ms (100.00%) | Memory: 8.8 MB (90.96%) - LeetSync

+++ 977-distinct-subsequences-ii/distinct-subsequences-ii.cpp
@@ -13,6 +13,6 @@ public:
             sum = (sum + total - count[idx] + MOD) % MOD;
             count[idx] = total;
         }
-        return sum;
+        return (int)sum;
     }
 };
